Fix help text and error IDs for aria-describedby (#746)
* Fix help text and error IDs for aria-describedby

Django recently started using `aria-describedby` to associate form
fields first with their help texts, and later also with error messages.

In #716, django-bootstrap5 was adapted to also include those the
necessary IDs in templates.

`aria-describedby` for fields can have different values:

-   `"{field.auto_id}_helptext"`
-   `"{field.auto_id}_error"`
-   both of them
-   a user defined value that is left unchanged

The approach in #716 was to use the value from `aria-describedby` for
the ID of the help text. Unfortunately, this only works in the first
case.

So I propose to solve this the same way the django templates solve this
and just hardcode the IDs.

* fix tests for django 5.2

* fix tests

---------

Co-authored-by: Dylan Verheul <dylan@dyve.net>